Change dotfile script.
It no longer adds dots before the new symlinks, so all the dotfiles have
to start with dots (src/bash/dotfiles/.bashrc instead of the old
src/bash/dotfiles/bashrc). This is better in case we ever have files we
want to symlink to that aren't dotfiles.

Some related changes were mistakenly added to this commit, for .vim
files, .vimrc, and maybe some others.